+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Added a `response` event with the HTTP status and response headers before body chunks on Android and iOS.
+
 ## [0.1.1] - 2026-08-27
 
 ### Fixed